Move float8 common_type to float8 header
